Add Keyring and KeyringClass types
The Keyring type represents the common API between all of our keyrings.
Currently the keyrings are written in JavaScript and they don't extend
a base type, but they are assumed to have certain methods and method
signatures.

The Keyring interface represents a keyring instance. It has mandatory
methods for basic functionality such as serialization/deserialization,
and getting and adding accounts. Some account management methods are
optional, like `exportAccount` which isn't possible to implement on
hardware wallets. The signing methods are all optional as well, as they
might not be implemented if the keyring doesn't support that type of
signing.

The KeyringClass type is used to validate that the class itself has the
expected constructor signature and static properties. This needs to be
a separate type because you can't validate constructor signatures or
static properties with an interface, and we rely upon both of these
things being consistent between all keyrings in the
`KeyringController`.

Note that this is just an attempt at capturing the existing keyring
API, without making any changes. We will be redesigning this API in the
near future.